♻️ Refactor get metadata function
